move http from util to adapter
