+++ cliboa/scenario/extract/http.py
@@ -15,9 +15,9 @@ import os
 
 from requests.auth import HTTPBasicAuth
 
+from cliboa.adapter.http import Download
 from cliboa.scenario.base import BaseStep
 from cliboa.scenario.validator import EssentialParameters
-from cliboa.util.http import Download
 
 
 class HttpExtract(BaseStep):
