+++ cliboa/test/adapter/test_http.py
@@ -3,7 +3,7 @@ import os
 import pytest
 from requests.exceptions import HTTPError
 
-from cliboa.util.http import Download, FormAuth
+from cliboa.adapter.http import Download, FormAuth
 
 
 class TestHttp(object):
